deploy/backend/requirements.txt
flask
pandas
openpyxl
flask-cors
requests
openai
python-dotenv
werkzeug
